refactor: 메인 파일 리팩토링
